docs: move illustrations to Illustrations
