+++ packages/components/src/Illustration/Scene/BrandMomentCaptureIntro/_docs/BrandMomentCaptureIntro.stickersheet.stories.tsx
@@ -7,7 +7,7 @@ import {
 import { BrandMomentCaptureIntro } from "../index"
 
 export default {
-  title: "Components/Illustrations/Scene/BrandMomentCaptureIntro",
+  title: "Illustrations/Scene/BrandMomentCaptureIntro",
   parameters: {
     chromatic: { disable: false, delay: 400 },
     controls: { disable: true },
