+++ packages/components/src/Illustration/Scene/BrandMomentCaptureIntro/_docs/BrandMomentCaptureIntro.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, StoryObj } from "@storybook/react"
 import { BrandMomentCaptureIntro } from "../index"
 
 const meta = {
-  title: "Components/Illustrations/Scene/BrandMomentCaptureIntro",
+  title: "Illustrations/Scene/BrandMomentCaptureIntro",
   component: BrandMomentCaptureIntro,
 } satisfies Meta<typeof BrandMomentCaptureIntro>
