+++ packages/components/src/Illustration/Scene/_docs/Scene.mdx
@@ -9,9 +9,7 @@ import * as SceneStories from "./Scene.stories"
 
 <ResourceLinks
   sourceCode="https://github.com/cultureamp/kaizen-design-system/tree/main/packages/components/src/Illustration/Scene"
-
 />
-
 
 <KAIOInstallation
   exportNames="BrandMomentPositiveOutro"
