+++ packages/components/src/Illustration/Scene/_docs/Scene.stickersheet.stories.tsx
@@ -54,7 +54,7 @@ import {
 } from "../index"
 
 export default {
-  title: "Components/Illustrations/Scene",
+  title: "Illustrations/Scene",
   parameters: {
     chromatic: { disable: false },
     controls: { disable: true },
